requirements.txt
pandas
torch
tqdm
transformers 
dask
pyarrow
fastparquet
fasttext
scikit-learn 
numpy
seaborn 
matplotlib
umap-learn
